==> house_prices/01_regression_model_ols_ridge_lasso_.ipynb <==
- Adding Dummy Variables/Handle Dummy Variable Trap
- Scale Data
- Stats OLS Model 
- Lasso
- Ridge

import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split 
from statsmodels.api import OLS
from sklearn.metrics import mean_squared_error

from sklearn.linear_model import LinearRegression, Lasso, Ridge

from scipy import stats

data = pd.read_pickle('data/train_no_missing_data.pkl')
data.info()

X = data.iloc[:, 1:-1]  # dropping Id column as its not a predictor 
y = data.iloc[:, -1].values

X.shape, y.shape

cat_columns = X.select_dtypes('category').columns
cat_columns , len(cat_columns)

## Adding Dummy Variables/Handle Dummy Variable Trap

def add_dummy(X, cat_columns):
    dropped_dummy_name =[]

    for col in cat_columns:
        col_dummy = pd.get_dummies(X[col], prefix=col)

        # drop last column to avoid dummy variable trap 
        dropped_dummy_name.append(col_dummy.columns[-1])
        #col_dummy = col_dummy.iloc[:,0:-1]            

        X = X.join(col_dummy)
        
    return X, dropped_dummy_name

X_new , dropped_dummy_name = add_dummy(X, cat_columns)
len(dropped_dummy_name)  

len(X_new.columns)

col_to_be_removed = []
col_to_be_removed.append(cat_columns)
col_to_be_removed.append(dropped_dummy_name)
col_to_be_removed = np.concatenate(col_to_be_removed)

len(col_to_be_removed)


# drop the columns for which dummy variables have been added
final_column = [col for col in X_new.columns if col not in col_to_be_removed]
len(final_column)

X_final = X_new[final_column].values

len(X_final[0,:])

## Scale Data

sc = StandardScaler()
X_scaled_final = sc.fit_transform(X_final)

## Stats OLS Model 

## Test/Train split 

X_scaled_final.shape

#Adding X_0 =1
X_ff = np.hstack((np.ones((X_scaled_final.shape[0], 1)), X_scaled_final ))
X_ff.shape

X_train, X_test, y_train, y_test = train_test_split(X_ff, y, test_size =0.2, random_state =0)

'''Fitting regressor on all variables'''
model = OLS(y_train,X_train)
regressor = model.fit()
regressor.summary()

y_pred = regressor.predict(X_test)

print('RMSE: ', np.sqrt(mean_squared_error(y_test, y_pred)))

regressor.save('data/ols_model.pkl')

## LASSO Model 

X_train, X_test, y_train, y_test = train_test_split(X_scaled_final, y, test_size =0.2, random_state =0)

lasso = Lasso(alpha= 0.5, max_iter=20000, random_state=0) # tune max_iter, alpha (regularization_parameter)

lasso.fit(X_train, y_train)

y_pred = lasso.predict(X_test)

print('RMSE: ', np.sqrt(mean_squared_error(y_test, y_pred)))

## Ridge Model

ridge = Ridge(alpha =0.5, random_state=0, max_iter=20000)

ridge.fit(X_train, y_train)

y_pred = ridge.predict(X_test)

print('RMSE: ', np.sqrt(mean_squared_error(y_test, y_pred)))

### check on test data

test_data = pd.read_csv('data/test.csv')
test_data.head()

col_with_na = [col for col in test_data.columns if test_data[col].isnull().sum() > 0]

for col in col_with_na:
    print('\n', col, ' <--> ', test_data[col].unique() , ' <--> ', test_data[col].isnull().sum())

test_X = test_data.iloc[:, 1:]  # dropping Id column as its not a predictor 

test_X_new , _ = add_dummy(test_X, cat_columns)


test_X.shape, test_X_new.shape

test_X


#test_X_final = test_X_new[final_column].values
