==> house_prices/04_random_forest.ipynb <==
- Load Data
- One-hot encoding of Categorical variables
- Test and train split
- RFR Model (Random Forest Regressor)
- Validation on Test Data

import pandas as pd 
import numpy as np 

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score


data = pd.read_pickle('data/train_data_cat_missing_binning_handled.pkl')
data.info()

data.describe()

data.select_dtypes('category').describe()

**Data** 
- 60 categorical columns
- 14 numerical 
    - exclude Id column
    - exclude target column 'SalePrice'
    
X = 60 categorical (with missing values)  + 12 numerical     
y = numerical 

data.columns

X = data[['MSSubClass', 'MSZoning', 'LotFrontage', 'LotArea', 'Street',
       'Alley', 'LotShape', 'LandContour', 'Utilities', 'LotConfig',
       'LandSlope', 'Neighborhood', 'Condition1', 'Condition2', 'BldgType',
       'HouseStyle', 'OverallQual', 'OverallCond', 'RoofStyle', 'RoofMatl',
       'Exterior1st', 'Exterior2nd', 'MasVnrType', 'MasVnrArea', 'ExterQual',
       'ExterCond', 'Foundation', 'BsmtQual', 'BsmtCond', 'BsmtExposure',
       'BsmtFinType1', 'BsmtFinSF1', 'BsmtFinType2', 'BsmtUnfSF',
       'TotalBsmtSF', 'Heating', 'HeatingQC', 'CentralAir', 'Electrical',
       '1stFlrSF', '2ndFlrSF', 'GrLivArea', 'BsmtFullBath', 'BsmtHalfBath',
       'FullBath', 'HalfBath', 'BedroomAbvGr', 'KitchenAbvGr', 'KitchenQual',
       'TotRmsAbvGrd', 'Functional', 'Fireplaces', 'FireplaceQu', 'GarageType',
       'GarageFinish', 'GarageCars', 'GarageArea', 'GarageQual', 'GarageCond',
       'PavedDrive', 'WoodDeckSF', 'OpenPorchSF', 'PoolQC', 'Fence',
       'MiscFeature', 'SaleType', 'SaleCondition', 'bin_yr_built',
       'bin_yr_remod', 'bin_garage_built', 'bin_yr_sold', 'bin_mo_sold']]
y = data['SalePrice']

X.shape, y.shape

## One-hot encoding for categorical v.


'''
    Convert given categorical columns to one-hot encoded columns
    drop 'NAN' column to avoid dummy variable trap
    drop categorical columns after 1-hot encoding 
'''

def get_one_hot_encoded_X(X, cat_colums):

    for col in cat_colums:
        temp = pd.get_dummies(X[col], prefix=col, dummy_na=True)
        final_col = [col for col in temp.columns if not col.endswith('nan')]
        X = X.join(temp[final_col])
    X = X.drop(cat_colums, axis =1)   
    return X  

cat_cols = X.select_dtypes('category').columns
X_new = get_one_hot_encoded_X(X, cat_cols)
X_new.head()

## Test and train split

X_train, X_test, y_train, y_test = train_test_split(X_new, y, test_size=0.2, random_state=0)

## Random Forest Model

model = RandomForestRegressor(n_estimators=20, random_state =0, verbose=2)

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

## Model Evaluation
- mse
- R^2

#r_2 = 1 - ssr/sst
r2_score(y_test,y_pred)

print( 'MSE: {0}, RMSE: {1}'.format(mean_squared_error(y_test, y_pred), np.sqrt(mean_squared_error(y_test, y_pred))))

pd.DataFrame({'y': y_test,
             'y_hat': y_pred,
             'mse': np.sqrt(mean_squared_error(y_test, y_pred))})

## Validation on Test Data

test_data = pd.read_pickle('data/test_data_cat_missing_binning_handled.pkl')
test_data.describe()


test_data.select_dtypes('category').describe()

# removing Id column 
X_test_data = test_data.iloc[:, 1: ]
X_test_data.head(2)

X_test_data_n = get_one_hot_encoded_X(X_test_data, cat_cols)
X_test_data_n.head(2)

# columns in model but not in  test data
missing_cols_in_test = [c for c in X_new.columns if c not in X_test_data_n.columns]
len(missing_cols_in_test)

missing_cols_in_test

# columns NOT in the model but are in test data
missing_cols_in_model = [c for c in X_test_data_n.columns if c not in X_new.columns]
len(missing_cols_in_model)

missing_cols_in_model

X_test_data_n.isnull().sum()

X_test_data_n.isnull().columns

#Let's fill nan in test data with 0:
# if its numeric field than 0 can work 
# if its one-hot-encoded then 0 means no value is there for thsi column
X_test_data_n.fillna(0, inplace = True)

X_test_data_n.isnull().sum().sum()

y_pred_t = model.predict(X_test_data_n)

### Join the prediction with Id and submit to kaggle

submission_df = pd.DataFrame({'Id': test_data['Id'], 
                             'SalePrice': y_pred_t})

submission_df

submission_df.to_csv('data/ah_submission_3.csv', index=False)
