==> house_prices/05_03_feature_selection.ipynb <==
# Feature Selection on House Prices Data

- L1 Regularization

import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel

%matplotlib inline

train_data = pd.read_pickle('data/final_train_data_feature_engineering_1hotenc_1.pkl')
train_data.info()

train_data.head()

train_data.describe()

X = train_data[train_data.columns.drop("SalePrice_log")]
y = train_data["SalePrice_log"]

X.shape, y.shape

X

## Normalize Data

sc = StandardScaler()

X.loc[: , X.columns] = sc.fit_transform(X.loc[:, X.columns])

## Test/Train split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

## Lasso Model

lasso = Lasso(alpha=0.02, random_state=0).fit(X_train, y_train)

model = SelectFromModel(lasso, prefit=True)

X_train_new = model.transform(X_train)
X_train_new.shape

X_train.shape

## Select  Features

selected_features = pd.DataFrame(model.inverse_transform(X_train_new), 
             index = X_train.index, 
             columns = X_train.columns)

selected_columns = selected_features.columns[selected_features.var()!=0]
selected_columns

## Save

final_cols = selected_columns.tolist()
final_cols.append('SalePrice_log')

pd.to_pickle(train_data[final_cols], 'data/final_train_data_l1_reg_v1.pkl')

## Checking Assumptions

Do it after selecting features 





### Linearity 

### Absence of Correlated errors

### Normality 

Check : 
- histogram and 
- normal probability plot; 
    - if no straight line, apply log transformation
    - if value zero, then can't apply log transformations
    

sns.distplot(train_data['SalePrice_log'])
fig = plt.figure()
res= stats.probplot(train_data['SalePrice_log'], plot=plt)

### Homoscedasticity 

- Between Sale Price and other numeric variable we must get a trending pattern (expecially after log transformations)
