==> house_prices/05_04_feature_transformation_test_data.ipynb <==
# Feature Transformation on House Prices Test Data

- Converting Column
    - Data Dictionary
- Feature Generation
    - Adding new features
    - Column Transformation
    - Binning 
- Categorical Encoding 
- Save data 


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 
import seaborn as sns
from pandas.plotting import scatter_matrix

from scipy import stats

%matplotlib inline

data = pd.read_csv('data/test.csv')
# data.info()

data.head()

data.describe()

data.select_dtypes('object').describe()

#final columns before dummification
final_cols = ['Id', 'MSSubClass',
 'MSZoning',
 'Street',
 'Alley',
 'LotShape',
 'LandContour',
 'Utilities',
 'LotConfig',
 'LandSlope',
 'Neighborhood',
 'Condition1',
 'Condition2',
 'BldgType',
 'HouseStyle',
 'OverallQual',
 'OverallCond',
 'RoofStyle',
 'RoofMatl',
 'Exterior1st',
 'Exterior2nd',
 'MasVnrType',
 'ExterQual',
 'ExterCond',
 'Foundation',
 'BsmtQual',
 'BsmtCond',
 'BsmtExposure',
 'BsmtFinType1',
 'BsmtFinType2',
 'BsmtUnfSF',
 'Heating',
 'HeatingQC',
 'CentralAir',
 'Electrical',
 '2ndFlrSF',
 'BsmtFullBath',
 'BsmtHalfBath',
 'FullBath',
 'HalfBath',
 'BedroomAbvGr',
 'KitchenAbvGr',
 'KitchenQual',
 'TotRmsAbvGrd',
 'Functional',
 'Fireplaces',
 'FireplaceQu',
 'GarageType',
 'GarageFinish',
 'GarageArea',
 'GarageQual',
 'GarageCond',
 'PavedDrive',
 'PoolQC',
 'Fence',
 'MiscFeature',
 'SaleType',
 'SaleCondition',
 'is_new_house',
 'is_recent_remolding',
 'basement_has_full_bathroom',
 'has_fireplace',
 'is_recent_garage_built',
 'has_garage',
 'has_pool',
 'LotFrontage_log',
 'LotArea_log',
 'MasVnrArea_log',
 'BsmtFinSF1_log',
 'BsmtFinSF2_log',
 'LowQualFinSF_log',
 'WoodDeckSF_log',
 'OpenPorchSF_log',
 'EnclosedPorch_log',
 '3SsnPorch_log',
 'ScreenPorch_log',
 'PoolArea_log',
 'MiscVal_log',
 'total_sqft_house_log',
 'bin_yr_built',
 'bin_yr_remod',
 'bin_yr_sold',
 'bin_mo_sold']

# EXtracted from feature selection process
final_feature_cols = ['Id', '2ndFlrSF', 'GarageArea', 'LotArea_log', 'OpenPorchSF_log',
       'EnclosedPorch_log', 'total_sqft_house_log', 'MSZoning_C (all)',
       'MSZoning_RM', 'Neighborhood_Crawfor', 'Condition1_Norm',
       'OverallQual_3', 'OverallQual_4', 'OverallQual_7', 'OverallQual_8',
       'OverallQual_9', 'OverallCond_3', 'OverallCond_4', 'ExterQual_TA',
       'Foundation_PConc', 'BsmtQual_TA', 'BsmtQual_Ex', 'BsmtExposure_No',
       'BsmtFinType1_GLQ', 'BsmtFinType1_Unf', 'HeatingQC_Ex', 'CentralAir_N',
       'FullBath_1', 'HalfBath_1', 'KitchenQual_TA', 'KitchenQual_Ex',
       'Functional_Typ', 'Fireplaces_2', 'GarageFinish_Fin', 'GarageCond_TA',
       'PavedDrive_Y', 'SaleCondition_Abnorml', 'has_fireplace_1',
       'bin_yr_remod_(1950, 1980]']


## Converting Column

### Data Dictionary

data_dict = {
    'MSSubClass': [20,30,40,45,50, 60,70,75,80,85,90,120,150,160,180,190], 
    'MSZoning': ['A', 'C (all)', 'FV', 'I', 'RH', 'RL', 'RP', 'RM'], 
    'Street': ['Grvl', 'Pave'], 
    'Alley': ['Grvl', 'Pave'], 
    'LotShape': ['Reg','IR1', 'IR2', 'IR3'], 
    'LandContour': ['Lvl','Bnk', 'HLS','Low'], 
    'Utilities': ['AllPub','NoSewr','NoSeWa','ELO'],
    'LotConfig': ['Inside', 'Corner', 'CulDSac', 'FR2', 'FR3'], 
    'LandSlope': ['Gtl', 'Mod', 'Sev'], 
    'Neighborhood': ['Blmngtn', 'Blueste', 'BrDale', 'BrkSide', 'ClearCr', 'CollgCr', 'Crawfor', 'Edwards', 'Gilbert'
                    ,'IDOTRR', 'MeadowV', 'Mitchel', 'Names', 'NoRidge', 'NPkVill', 'NridgHt', 'NWAmes', 'OldTown', 
                    'SWISU', 'Sawyer', 'SawyerW', 'Somerst', 'StoneBr', 'Timber', 'Veenker'], 
    'Condition1': ['Artery', 'Feedr', 'Norm', 'RRNn', 'RRAn', 'PosN', 'PosA', 'RRNe', 'RRAe'], 
    'Condition2': ['Artery', 'Feedr', 'Norm', 'RRNn', 'RRAn', 'PosN', 'PosA', 'RRNe', 'RRAe'], 
    'BldgType': ['1Fam', '2FmCon', 'Duplx', 'TwnhsE', 'TwnhsI'], 
    'HouseStyle': ['1Story', '1.5Fin', '1.5Unf', '2Story', '2.5Fin', '2.5Unf', 'SFoyer', 'SLvl'], 
    'OverallQual': np.arange(1,11), 
    'OverallCond': np.arange(1,11), 
    'RoofStyle': ['Flat', 'Gable', 'Gambrel', 'Hip', 'Mansard', 'Shed'], 
    'RoofMatl': ['ClyTile', 'CompShg', 'Membran', 'Metal', 'Roll', 'Tar&Grv', 'WdShake', 'WdShngl'], 
    'Exterior1st': ['AsbShng', 'AsphShn', 'BrkComm', 'BrkFace', 'CBlock', 'CemntBd', 'HdBoard', 'ImStucc', 'MetalSd', 
                   'Other', 'Plywood', 'PreCast', 'Stone', 'Stucco', 'VinylSd', 'Wd Sdng', 'WdShing'],
    'Exterior2nd': ['AsbShng', 'AsphShn', 'BrkComm', 'BrkFace', 'CBlock', 'CemntBd', 'HdBoard', 'ImStucc', 'MetalSd', 
                   'Other', 'Plywood', 'PreCast', 'Stone', 'Stucco', 'VinylSd', 'Wd Sdng', 'WdShing'],
    'MasVnrType': ['BrkCmn', 'BrkFace', 'CBlock', 'None', 'Stone'], 
    'ExterQual': ['Po', 'Fa', 'TA', 'Gd','Ex'], 
    'ExterCond': ['Po', 'Fa', 'TA', 'Gd','Ex'], 
    'Foundation': ['BrkTil', 'CBlock', 'PConc', 'Slab', 'Stone', 'Wood'], 
    'BsmtQual': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'BsmtCond': ['Po', 'Fa', 'TA', 'Gd','Ex'], 
    'BsmtExposure': ['No', 'Mn', 'Av', 'Gd'], 
    'BsmtFinType1': ['GLQ', 'ALQ', 'BLQ', 'Rec', 'LwQ', 'Unf'],
    'BsmtFinType2': ['GLQ', 'ALQ', 'BLQ', 'Rec', 'LwQ', 'Unf'],
    'Heating': ['Floor', 'GasA', 'GasW', 'Grav', 'OthW', 'Wall'], 
    'HeatingQC': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'CentralAir': ['N', 'Y'],
    'Electrical': ['SBrkr', 'FuseA', 'FuseF', 'FuseP', 'Mix'], 
    'KitchenQual': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'Functional': ['Typ', 'Min1', 'Min2', 'Mod', 'Maj1', 'Maj2', 'Sev', 'Sal'], 
    'FireplaceQu': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'GarageType': ['2Types', 'Attchd', 'Basment', 'BuiltIn', 'CarPort', 'Detchd'],
    'GarageFinish': ['Fin', 'RFn', 'Unf'],
    'GarageQual': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'GarageCond': ['Po', 'Fa', 'TA', 'Gd','Ex'],
    'PavedDrive': ['N', 'P', 'Y'], 
    'PoolQC':  ['Fa', 'TA', 'Gd','Ex'], 
    'Fence': ['GdPrv', 'MnPrv', 'GdWo', 'MnWw'], 
    'MiscFeature': ['Elev', 'Gar2', 'Othr', 'Shed', 'TenC'],
    'SaleType': ['WD', 'CWD', 'VWD', 'New', 'COD', 'Con', 'ConLw', 'ConLI', 'ConLD','Oth'], 
    'SaleCondition': ['Normal', 'Abnorml', 'AdjLand', 'Alloca', 'Family', 'Partial'], 
    'BsmtFullBath': np.arange(0,4), 
    'BsmtHalfBath': np.arange(0, 3),
    'FullBath': np.arange(0,5),
    'HalfBath': np.arange(0,3), 
    'BedroomAbvGr': np.arange(0,9),
    'KitchenAbvGr' : np.arange(0,4), 
    'TotRmsAbvGrd' : np.arange(0, 16), 
    'Fireplaces' : np.arange(0,5), 
    'GarageCars' : np.arange(0,6), 
    'MoSold' : np.arange(1,13)
   
}

order_cat = ['OverallQual', 'OverallCond', 'ExterQual', 'ExterCond', 
             'BsmtQual', 'BsmtCond', 'BsmtExposure', 'BsmtFinType1','BsmtFinType2', 
             'BsmtFullBath', 'BsmtHalfBath',
             'HeatingQC', 'CentralAir',
             'KitchenQual','KitchenAbvGr','Kitchen'
             'FireplaceQu', 'Fireplaces',
             'GarageQual', 'GarageCond', 'GarageCars', 'GarageFinish'
             'PoolQC',  
             'FullBath', 'HalfBath', 
             'BedroomAbvGr', 'TotRmsAbvGrd',  'Bedroom']

# year is categorical variable becuase it makes no sense to add two years. 
# Also house build in 2010 is better than the one built in 1980

year_cols = ['YearBuilt', 'YearRemodAdd', 'GarageYrBlt', 'YrSold', 'MoSold'] 

len(data_dict.keys()) , len(order_cat), len(year_cols)

cat_cols = data.select_dtypes('object').columns.values

# columns that are categorical but not object type
additional_cat = [key for key in data_dict.keys() if key not in cat_cols]
cat_cols = np.concatenate((cat_cols, additional_cat))
cat_cols

## convert type and also add order 
for col in cat_cols:
    is_ordered = True if col in order_cat else False
    data[col] = pd.Categorical(data[col], categories=data_dict[col], ordered=is_ordered)

# add categories for year columns
for col in year_cols:
    data[col] = pd.Categorical(data[col], ordered=True)
    
data.select_dtypes('category').describe()    

## Adding new features


def new_column_mapping(data):
    
    data['is_new_house'] = np.where(data['YearBuilt'] > 2000 , 1, 0)
    data['is_recent_remolding'] = np.where(data['YearRemodAdd'] > 2000 , 1, 0)
    data['basement_has_full_bathroom'] = np.where(data['BsmtFullBath'] > 0, 1, 0) 
    data['has_fireplace'] = np.where(data['Fireplaces'] > 0, 1, 0) 
    data['is_recent_garage_built'] = np.where(data['GarageYrBlt'] > 2000 , 1, 0)
    data['has_garage'] = np.where(data['GarageYrBlt'].isnull() , 0, 1)
    data['has_pool'] = np.where(data['PoolArea'] > 0, 1, 0)

    garage_area = np.where(data['GarageType'].isin(['Attchd', 'Basment', 'BuiltIn']), data['GarageArea'], 0)
    data['total_sqft_house'] = data['GrLivArea'] + garage_area + data['EnclosedPorch'] + data['TotalBsmtSF']

    # convert to categories
    for c in ['is_new_house', 'is_recent_remolding', 'basement_has_full_bathroom', 'has_fireplace',
                      'is_recent_garage_built', 'has_garage', 'has_pool']:
        data[c] = pd.Categorical(data[c])

# data = 
new_column_mapping(data)

## Column Transformation


to_apply_transformation = ['LotFrontage', 'LotArea', 'MasVnrArea', 'BsmtFinSF1', 'BsmtFinSF2', 
                           'TotalBsmtSF', '1stFlrSF', 'LowQualFinSF', 'GrLivArea',  
                           'WoodDeckSF', 'OpenPorchSF', 'EnclosedPorch', '3SsnPorch', 'ScreenPorch', 
                           'PoolArea', 'MiscVal','total_sqft_house']

# to_apply_transformation_has_zero =  ['MasVnrArea', 'BsmtFinSF1', 'BsmtFinSF2', 'TotalBsmtSF', 
#                                      'LowQualFinSF', 'BsmtHalfBath', 'KitchenAbvGr', 'WoodDeckSF', 
#                                      'OpenPorchSF', 'EnclosedPorch', '3SsnPorch', 'ScreenPorch', 
#                                      'PoolArea', 'MiscVal']


def apply_transformation(data):
    
    for col in to_apply_transformation:

        col_name = col + '_log'
        if (data[col] < 0).sum() <= 0: ## to avoid log failure
            data[col_name] = np.log(data[col]+ 0.0001)
        else:    
            data[col_name] = np.log(data[col])

#     return data

# data = 
apply_transformation(data)

## Binning 


def binning(data):
    yr_built_bins = [1850, 1950, 1975, 2000, 2020]
    data['bin_yr_built'] = pd.cut(data['YearBuilt'], bins= yr_built_bins, precision=0)

    yr_remod_bins = [1950, 1980, 1990, 2000, 2020]
    data['bin_yr_remod'] = pd.cut(data['YearRemodAdd'], bins= yr_remod_bins, precision=0)

    yr_garage_bins = [1880, 1950, 1980, 1990, 2000,2020]
    data['bin_garage_built'] = pd.cut(data['GarageYrBlt'], bins= yr_garage_bins, precision=0)

    yr_sold_bins = [2006, 2007, 2008, 2009, 2010]
    data['bin_yr_sold'] = pd.cut(data['YrSold'], bins= yr_sold_bins, precision=0)

    month_sold_bins = [1, 4, 6, 9, 12]
    #month_sold_bins_lbl = ['spring', 'summer', 'fall', 'winter']
    data['bin_mo_sold'] = pd.cut(data['MoSold'], bins= month_sold_bins, precision=0)

#     return data
    

# data =
binning(data)

## Categorical Encoding 
- One-hot Encoding of all categorical variables with na dropped and kept 


'''
    Convert given categorical columns to one-hot encoded columns
    drop 'NAN' column to avoid dummy variable trap
    drop categorical columns after 1-hot encoding 
    
    dummy_trap = True --> delete 1 dummy variable; 
        if column has null values then delete NA 
        if column has 1/0 value then delete 0 column
        
        Na and 0 signify lack of something 
'''

def get_one_hot_encoded_X(X, cat_colums, dummy_trap=True, delete_original_cat= True):

    for col in cat_colums:
        
        temp = pd.get_dummies(X[col], prefix=col, dummy_na= False if data[col].isnull().sum() == 0 else True)

        col_dummies = temp.columns
        
        if dummy_trap:
            col_dummies = [c for c in col_dummies if not( c.endswith('0') or c.endswith('nan'))]
        
        X = X.join(temp[col_dummies])

    if delete_original_cat:    
        X = X.drop(cat_colums, axis =1)   
        
    return X  

cat_cols_final = data[final_cols].select_dtypes('category').columns

X_new = get_one_hot_encoded_X(data[final_cols], cat_cols_final, dummy_trap= True, delete_original_cat= True)
final_cols_with_dummies = X_new.columns
X_new.shape

transformed_test_data = X_new[final_feature_cols]

## NAN in Numerical Columns

fill with 0

transformed_test_data.isnull().sum().sort_values(ascending=False)

transformed_test_data['GarageArea'].fillna(0, inplace=True)
transformed_test_data['total_sqft_house_log'].fillna(0, inplace=True)

transformed_test_data.isnull().sum()

# Save 

data.shape, data[final_cols].shape, transformed_test_data.shape

#save data to file

pd.to_pickle(transformed_test_data, 'data/data_test_l1_1.pkl')
# pd.to_pickle(data[final_cols], 'data/final_data_feature_engineering_1.pkl')
# pd.to_pickle(X_new, 'data/final_data_feature_engineering_1hotenc_1.pkl')

train_data = pd.read_pickle('data/final_train_data_feature_engineering_1hotenc_1.pkl')
# train_data.info()

# columns missing in test
len(set(train_data.columns) - set(X_new.columns))

len(set(X_new.columns) - set(train_data.columns))

len(train_data.columns), len(X_new.columns)

set(train_data.columns) - set(X_new.columns)

Right now, these are not part of final feature column, so it does not matter
