==> house_prices/05_05_random_forest_l1.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score 


data = pd.read_pickle('data/final_train_data_l1_reg_v1.pkl')
data.shape

data.head()

X = data.iloc[:, :-1]
y = data.iloc[:, -1]

X.shape, y.shape

X.head()

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

model = RandomForestRegressor()

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

r2_score(y_test, y_pred)

print( 'MSE: {0}, RMSE: {1}'.format(mean_squared_error(y_test, y_pred), np.sqrt(mean_squared_error(y_test, y_pred))))

## Validation on Test Data

- need to apply transformation 
- select the final cols 

test_data = pd.read_pickle('data/data_test_l1_1.pkl')
test_data.head()

test_data.isnull().sum().sum()

test_X = test_data.iloc[:, 1:]
Id_col = test_data.iloc[:, 0]

y_pred_test = model.predict(test_X)

submission = pd.DataFrame({'Id': Id_col, 
             'SalePrice': y_pred_test})
submission.head()

submission.to_csv('data/ah_submission_4.csv', index=False)
