==> house_prices/06_02_feature_transformation_test_data.ipynb <==
# Feature Engineering on House Prices Data 

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 
import seaborn as sns
from pandas.plotting import scatter_matrix

from scipy import stats

%matplotlib inline

test_data = pd.read_csv('data/test.csv')
# train_data.info()

test_data.head()

test_data.describe()

test_data.select_dtypes('object').describe()

## Converting Column

### Data Dictionary

data_dict =  pd.read_pickle('data/data_dict_06_v1.pkl')

order_cat = ['OverallQual', 'OverallCond', 'ExterQual', 'ExterCond', 
             'BsmtQual', 'BsmtCond', 'BsmtExposure', 'BsmtFinType1','BsmtFinType2', 
             'HeatingQC', 'CentralAir', 'KitchenQual','FireplaceQu',
             'GarageQual', 'GarageCond', 'GarageCars', 'GarageFinish'
             'PoolQC']

# year is categorical variable becuase it makes no sense to add two years. 
# Also house build in 2010 is better than the one built in 1980

year_cols = ['YearBuilt', 'YearRemodAdd', 'GarageYrBlt', 'YrSold', 'MoSold'] 

len(data_dict.keys()) , len(order_cat), len(year_cols)

cat_cols = test_data.select_dtypes('object').columns.values

# columns that are categorical but not object type
additional_cat = [key for key in data_dict.keys() if key not in cat_cols]
cat_cols = np.concatenate((cat_cols, additional_cat))
cat_cols

## convert type and also add order 
for col in cat_cols:
    is_ordered = True if col in order_cat else False
    test_data[col] = pd.Categorical(test_data[col], categories=data_dict[col], ordered=is_ordered)

# add categories for year columns
for col in year_cols:
    test_data[col] = pd.Categorical(test_data[col], ordered=True)
    
test_data.select_dtypes('category').describe()    

## Adding new features

test_data['is_new_house'] = np.where(test_data['YearBuilt'] > 2000 , 1, 0)
test_data['is_recent_remolding'] = np.where(test_data['YearRemodAdd'] > 2000 , 1, 0)
test_data['basement_has_full_bathroom'] = np.where(test_data['BsmtFullBath'].cat.codes > 0, 1, 0) 
test_data['has_fireplace'] = np.where(test_data['Fireplaces'].cat.codes > 0, 1, 0) 
test_data['is_recent_garage_built'] = np.where(test_data['GarageYrBlt'] > 2000 , 1, 0)
test_data['has_garage'] = np.where(test_data['GarageYrBlt'].isnull() , 0, 1)
test_data['has_pool'] = np.where(test_data['PoolArea'] > 0, 1, 0)

# convert to categories
for c in ['is_new_house', 'is_recent_remolding', 'basement_has_full_bathroom', 'has_fireplace',
                  'is_recent_garage_built', 'has_garage', 'has_pool']:
    test_data[c] = pd.Categorical(test_data[c])


## Binning 
- Equal distribution binning 
- [ --> includes  and ( --> excludes

yr_built_bins = [1872, 1950, 1975, 2000, 2025]
yr_built_bins_interval = pd.IntervalIndex.from_breaks(yr_built_bins, closed ='left')
test_data['bin_yr_built'] = pd.cut(test_data['YearBuilt'], bins=yr_built_bins_interval, precision=0)

yr_remod_bins = [1950, 1980, 1990, 2000, 2025]
yr_remod_bins_interval = pd.IntervalIndex.from_breaks(yr_remod_bins, closed ='left')
test_data['bin_yr_remod'] = pd.cut(test_data['YearRemodAdd'], bins=yr_remod_bins_interval, precision=0)

yr_garage_bins = [1872, 1950, 1980, 1990, 2000,2025]
yr_garage_bins_interval = pd.IntervalIndex.from_breaks(yr_garage_bins, closed ='left')
test_data['bin_garage_built'] = pd.cut(test_data['GarageYrBlt'], bins=yr_garage_bins_interval, precision=0)

yr_sold_bins = [2006, 2007, 2008, 2009, 2025]
yr_sold_bins_interval = pd.IntervalIndex.from_breaks(yr_sold_bins, closed ='left')
test_data['bin_yr_sold'] = pd.cut(test_data['YrSold'], bins=yr_sold_bins_interval , precision=0)

month_sold_bins = [1, 4, 6, 9, 13]
month_sold_bins_interval = pd.IntervalIndex.from_breaks(month_sold_bins, closed ='left')
test_data['bin_mo_sold'] = pd.cut(test_data['MoSold'], bins=month_sold_bins_interval, precision=0)


## Fill missing values

Only for Regression Data

# test_regression_data = test_data
# test_tree_data = test_data.copy()

test_data['MSZoning'].fillna("RL", inplace=True)
test_data['Utilities'].fillna("AllPub", inplace=True)
test_data['Exterior1st'].fillna("VinylSd", inplace=True)
test_data['Exterior2nd'].fillna("VinylSd", inplace=True)
test_data['BsmtFullBath'].fillna(0, inplace=True)
test_data['BsmtHalfBath'].fillna(0, inplace=True)
test_data['KitchenQual'].fillna('TA', inplace=True)
test_data['Functional'].fillna('Typ', inplace=True)
test_data['GarageYrBlt'] = np.where( test_data['GarageYrBlt'].isnull(), test_data['YearBuilt'], test_data['GarageYrBlt'])
test_data['GarageCars'].fillna(2, inplace=True)
test_data['SaleType'].fillna('WD', inplace=True)
test_data['bin_garage_built'].fillna(test_data['bin_garage_built'].mode()[0], inplace=True)


for col in test_data.select_dtypes(include='category').columns:
    if test_data[col].isnull().sum() > 0:
        print(col)
        test_data[col].fillna("None", inplace=True)

for col in test_data.select_dtypes(exclude='category').columns:
    if test_data[col].isnull().sum() > 0:
        print(col)
        test_data[col].fillna(0, inplace=True)

test_data.isnull().sum().sort_values(ascending=False)

**For Tree Models**

Didn't work

# test_tree_data['PoolQC'].fillna('None', inplace=True)
# test_tree_data['MiscFeature'].fillna('None', inplace=True)
# test_tree_data['Alley'].fillna('None', inplace=True)
# test_tree_data['Fence'].fillna('None', inplace=True)
# test_tree_data['FireplaceQu'].fillna('None', inplace=True)
# test_tree_data['GarageCond'].fillna('None', inplace=True)
# test_tree_data['GarageType'].fillna('None', inplace=True)

# # houses who have no garage; so year really does not matter. Set the year to the year the house was built
# test_tree_data['GarageYrBlt'] = np.where( test_tree_data['GarageYrBlt'].isnull(), test_tree_data['YearBuilt'], test_tree_data['GarageYrBlt'])

# test_tree_data['GarageFinish'].fillna('None', inplace=True)
# test_tree_data['GarageQual'].fillna('None', inplace=True)

# test_tree_data['BsmtExposure'].fillna('None', inplace=True)
# test_tree_data['BsmtFinType2'].fillna('None', inplace=True)
# test_tree_data['BsmtFinType1'].fillna('None', inplace=True) 
# test_tree_data['BsmtCond'].fillna('None', inplace=True)
# test_tree_data['BsmtQual'].fillna('None', inplace=True)
# test_tree_data['MasVnrType'].fillna('None', inplace=True) 


## Categorical Encoding 
- One-hot Encoding of all categorical variables with na dropped and kept 


'''
    Convert given categorical columns to one-hot encoded columns
    drop 'NAN' column to avoid dummy variable trap
    drop categorical columns after 1-hot encoding 
    
    dummy_trap = True --> delete the first dummy variable, incase a column has 'None', it'll be the first column; 
        if column has null values then delete NA 
        if column has 1/0 value then delete 0 column
        
        Na and 0 signify lack of something 
'''

def get_one_hot_encoded_X(X, cat_colums, dummy_trap=True, delete_original_cat= True):

    for col in cat_colums:
        
        temp = pd.get_dummies(X[col], prefix=col, dummy_na= False, drop_first=dummy_trap)
        X = X.join(temp)

    if delete_original_cat:    
        X = X.drop(cat_colums, axis =1)   
        
    return X  

final_cols = pd.read_pickle('data/final_cols_06_v1.pkl')
final_cols.remove('SalePrice')
final_cols.append('Id')

test = test_data[final_cols]
test.shape, test.isnull().sum().sum()

cat_cols_final = test_data[final_cols].select_dtypes('category').columns
cat_cols_final.shape

test_regression_data = get_one_hot_encoded_X(test, cat_cols_final, dummy_trap= True, delete_original_cat= True)
test_tree_data = get_one_hot_encoded_X(test, cat_cols_final, dummy_trap= False, delete_original_cat= True)

test.shape, test_regression_data.shape, test_tree_data.shape

## Missing check for Tree Model

missing = pd.DataFrame(test_tree_data.isnull().sum().sort_values(ascending=False))
missing[missing[0] > 0 ]

# test_tree_data['LotFrontage'].fillna(0, inplace=True)
# test_tree_data['MasVnrArea'].fillna(0, inplace=True)
# test_tree_data['BsmtFinSF2'].fillna(0, inplace=True) 
# test_tree_data['TotalBsmtSF'].fillna(0, inplace=True)
# test_tree_data['BsmtUnfSF'].fillna(0, inplace=True)
# test_tree_data['GarageArea'].fillna(0, inplace=True) 
# test_tree_data['BsmtFinSF1'].fillna(0, inplace=True) 

missing = pd.DataFrame(test.isnull().sum().sort_values(ascending=False))
missing[missing[0] > 0 ]


# for col in missing[missing[0] > 0 ].index:
#     if test[col].dtype == np.float64:
#         test[col].fillna(0, inplace=True)
#     else:
#         test[col].fillna('NA', inplace=True)
    

test['GarageArea'].dtype == np.float64

# Save 

#save data to file

pd.to_pickle(test_regression_data, 'data/test_regression_data_06_v1.pkl')
pd.to_pickle(test_tree_data, 'data/test_tree_data_06_v1.pkl')
pd.to_pickle(test, 'data/test_catboost_data_06_v1.pkl')
